* 'master' of https://github.com/ajanata/PretendYoureXyzzy: (273 commits)
  Fix JS error because of prefix
  update postgresql JDBC driver to latest version
  big link to server list page
  changed geoip version to 2.8.1 to make build work
  Added min/default/max values to js config
  Moved min/default/max player limit, spectator limit, score limit and blank card limit to pyx.properties
  Disable inactive card sets by default.
  update changelog and privacy text
  Allow admins to spectate any password protected game
  Adding Docker docs to README
  Add more info to stats.jsp to allow server lists to disambiguate features
  Allow admins to join any game
  Add option to disable game chat.
  Add option to disable fill-in-the-blank cards entirely.
  Fix for #174 (#176)
  Check user address against X-Forwarded-For
  Bugfixes and nick filter.
  changelog updates
  Show more permalinks to users.
  Send global chat enabled with FirstLoad
  ...

# Conflicts:
#	.classpath
#	.gitignore
#	WebContent/addcard.jsp
#	WebContent/admin.jsp
#	WebContent/cardsets.jsp
#	WebContent/game.jsp
#	WebContent/index.jsp
#	WebContent/js/cah.app.js
#	WebContent/js/cah.constants.js
#	WebContent/stats.jsp
#	pom.xml
#	src/hibernate.cfg.xml
#	src/main/java/net/socialgamer/cah/Constants.java
#	src/main/java/net/socialgamer/cah/handlers/Handlers.java
#	src/net/socialgamer/cah/handlers/RegisterHandler.java
#	src/net/socialgamer/cah/servlets/JavascriptConfigServlet.java

+++ WebContent/cardsets.jsp
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <%--
-Copyright (c) 2012, Andy Janata
+Copyright (c) 2012-2018, Andy Janata
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted
@@ -27,22 +27,28 @@ Administration tools.
 @author Andy Janata (ajanata@socialgamer.net)
 --%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
+<%@ page import="com.google.inject.Injector" %>
+<%@ page import="com.google.inject.Key" %>
+<%@ page import="com.google.inject.TypeLiteral" %>
 <%@ page import="java.util.ArrayList" %>
 <%@ page import="java.util.List" %>
+<%@ page import="java.util.Set" %>
+<%@ page import="net.socialgamer.cah.CahModule.Admins" %>
 <%@ page import="net.socialgamer.cah.HibernateUtil" %>
-<%@ page import="net.socialgamer.cah.db.BlackCard" %>
-<%@ page import="net.socialgamer.cah.db.CardSet" %>
-<%@ page import="net.socialgamer.cah.db.WhiteCard" %>
+<%@ page import="net.socialgamer.cah.StartupUtils" %>
+<%@ page import="net.socialgamer.cah.db.PyxBlackCard" %>
+<%@ page import="net.socialgamer.cah.db.PyxCardSet" %>
+<%@ page import="net.socialgamer.cah.db.PyxWhiteCard" %>
 <%@ page import="net.socialgamer.cah.RequestWrapper" %>
 <%@ page import="org.apache.commons.lang3.StringEscapeUtils" %>
 <%@ page import="org.hibernate.Session" %>
 <%@ page import="org.hibernate.Transaction" %>
 <%
 RequestWrapper wrapper = new RequestWrapper(request);
-String remoteAddr = wrapper.getRemoteAddr();
-//TODO better access control than hard-coding IP addresses.
-if (!(remoteAddr.equals("0:0:0:0:0:0:0:1") || remoteAddr.equals("127.0.0.1") ||
-    remoteAddr.startsWith("10.") || remoteAddr.equals("98.210.81.226"))) {
+ServletContext servletContext = pageContext.getServletContext();
+Injector injector = (Injector) servletContext.getAttribute(StartupUtils.INJECTOR);
+Set<String> admins = injector.getInstance(Key.get(new TypeLiteral<Set<String>>(){}, Admins.class));
+if (!admins.contains(wrapper.getRemoteAddr())) {
   response.sendError(403, "Access is restricted to known hosts");
   return;
 }
@@ -54,10 +60,10 @@ Session hibernateSession = HibernateUtil.instance.sessionFactory.openSession();
 // cheap way to make sure we can close the hibernate session at the end of the page
 try {
   String editParam = request.getParameter("edit");
-  CardSet editCardSet = null;
+  PyxCardSet editCardSet = null;
   if (null != editParam) {
     try {
-      editCardSet = (CardSet)hibernateSession.load(CardSet.class, Integer.parseInt(editParam));
+      editCardSet = (PyxCardSet)hibernateSession.load(PyxCardSet.class, Integer.parseInt(editParam));
     } catch (NumberFormatException nfe) {
       messages.add("Unable to parse or locate requested card set to edit.");
     }
@@ -66,7 +72,7 @@ try {
   String deleteParam = request.getParameter("delete");
   if (null != deleteParam) {
     try {
-      editCardSet = (CardSet)hibernateSession.load(CardSet.class, Integer.parseInt(deleteParam));
+      editCardSet = (PyxCardSet)hibernateSession.load(PyxCardSet.class, Integer.parseInt(deleteParam));
       Transaction t = hibernateSession.beginTransaction();
       hibernateSession.delete(editCardSet);
       t.commit();
@@ -85,17 +91,27 @@ try {
     try {
       id = Integer.parseInt(idParam);
       if (-1 == id) {
-        editCardSet = new CardSet();
+        editCardSet = new PyxCardSet();
       } else {
-        editCardSet = (CardSet)hibernateSession.load(CardSet.class, id);
+        editCardSet = (PyxCardSet)hibernateSession.load(PyxCardSet.class, id);
       }
       if (null != editCardSet) {
         String nameParam = request.getParameter("cardSetName");
+        String descriptionParam = request.getParameter("cardSetDescription");
+        String weightParam = request.getParameter("cardSetWeight");
         String activeParam = request.getParameter("active");
         String baseDeckParam = request.getParameter("baseDeck");
         String[] selectedBlackCardsParam = request.getParameterValues("selectedBlackCards");
         String[] selectedWhiteCardsParam = request.getParameterValues("selectedWhiteCards");
-        if (null == nameParam || nameParam.isEmpty() || null == selectedBlackCardsParam ||
+        int weight = -1;
+        try {
+          weight = Integer.valueOf(weightParam);
+        } catch (Exception e) {
+          // pass
+        }
+        if (weight <= 0 || weight > 9999) {
+          messages.add("Weight must be a positive integer less than 10000.");
+        } else if (null == nameParam || nameParam.isEmpty() || null == selectedBlackCardsParam ||
             null == selectedWhiteCardsParam) {
           messages.add("You didn't specify something.");
           if (-1 == id) {
@@ -103,6 +119,8 @@ try {
           }
         } else {
           editCardSet.setName(nameParam);
+          editCardSet.setDescription(descriptionParam);
+          editCardSet.setWeight(weight);
           editCardSet.setActive("on".equals(activeParam));
           editCardSet.setBaseDeck("on".equals(baseDeckParam));
           List<Integer> blackCardIds = new ArrayList<Integer>(selectedBlackCardsParam.length);
@@ -114,12 +132,12 @@ try {
             whiteCardIds.add(Integer.parseInt(wc));
           }
           @SuppressWarnings("unchecked")
-          List<BlackCard> realBlackCards = hibernateSession.createQuery(
-              "from BlackCard where id in (:ids)").setParameterList("ids", blackCardIds).
+          List<PyxBlackCard> realBlackCards = hibernateSession.createQuery(
+              "from PyxBlackCard where id in (:ids)").setParameterList("ids", blackCardIds).
               setReadOnly(true).list();
           @SuppressWarnings("unchecked")
-          List<WhiteCard> realWhiteCards = hibernateSession.createQuery(
-              "from WhiteCard where id in (:ids)").setParameterList("ids", whiteCardIds).
+          List<PyxWhiteCard> realWhiteCards = hibernateSession.createQuery(
+              "from PyxWhiteCard where id in (:ids)").setParameterList("ids", whiteCardIds).
               setReadOnly(true).list();
           editCardSet.getBlackCards().clear();
           editCardSet.getBlackCards().addAll(realBlackCards);
@@ -141,17 +159,16 @@ try {
   }
   
   @SuppressWarnings("unchecked")
-  List<CardSet> cardSets = hibernateSession.createQuery("from CardSet order by id")
+  List<PyxCardSet> cardSets = hibernateSession.createQuery("from PyxCardSet order by weight, name")
       .setReadOnly(true).list();
   
   @SuppressWarnings("unchecked")
-  List<BlackCard> blackCards = hibernateSession.createQuery("from BlackCard order by id")
+  List<PyxBlackCard> blackCards = hibernateSession.createQuery("from PyxBlackCard order by id")
       .setReadOnly(true).list();
   
   @SuppressWarnings("unchecked")
-  List<WhiteCard> whiteCards = hibernateSession.createQuery("from WhiteCard order by id")
+  List<PyxWhiteCard> whiteCards = hibernateSession.createQuery("from PyxWhiteCard order by id")
       .setReadOnly(true).list();
-
 %>
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
     "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
@@ -159,7 +176,7 @@ try {
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
 <title>PYX - Edit Card Sets</title>
-<script type="text/javascript" src="js/jquery-1.8.2.js"></script>
+<script type="text/javascript" src="js/jquery-1.11.3.min.js"></script>
 <script type="text/javascript">
   $(document).ready(function() {
     $('#addBlackCards').click(function() {
@@ -218,9 +235,13 @@ select {
 </style>
 </head>
 <body>
-<% for (String message : messages) { %>
-  <h3><%= message %></h3>
-<% } %>
+<%
+  for (String message : messages) {
+%>
+  <h3><%=message%></h3>
+<%
+  }
+%>
 <h2>Existing card sets</h2>
 <table style="1px solid black">
   <thead>
@@ -228,48 +249,80 @@ select {
       <th>Name</th>
       <th>Delete</th>
       <th>Edit</th>
+      <th>Weight</th>
+      <th>Blacks</th>
+      <th>Whites</th>
+      <th>Active</th>
     </tr>
   </thead>
   <tbody>
-    <% for (CardSet cardSet : cardSets) { %>
+    <%
+      for (PyxCardSet cardSet : cardSets) {
+    %>
       <tr>
-        <td><%= StringEscapeUtils.escapeXml(cardSet.getName()) %></td>
-        <td><a href="?delete=<%= cardSet.getId() %>">Delete</a></td>
-        <td><a href="?edit=<%= cardSet.getId() %>">Edit</a></td>
+        <td><%=cardSet.getName()%></td>
+        <td><a href="?delete=<%=cardSet.getId()%>" onclick="return confirm('Are you sure?')">Delete</a></td>
+        <td><a href="?edit=<%=cardSet.getId()%>">Edit</a></td>
+        <td><%=cardSet.getWeight()%></td>
+        <td><%=cardSet.getBlackCards().size()%></td>
+        <td><%=cardSet.getWhiteCards().size()%></td>
+        <td><%=cardSet.isActive()%></td>
       </tr>
-    <% } %>
+    <%
+      }
+    %>
   </tbody>
 </table>
 <a href="cardsets.jsp">Create New</a>
 <form action="cardsets.jsp" method="post" id="editForm">
   <input type="hidden" name="action" value="edit" />
   <input type="hidden" name="cardSetId"
-      value="<%= editCardSet != null ? editCardSet.getId() : -1 %>" />
+      value="<%=editCardSet != null ? editCardSet.getId() : -1%>" />
   <h2>
-    <% if (editCardSet != null) { %>
-      Editing <span style="text-decoration:italic"><%= editCardSet.getName() %></span>
-    <% } else { %>
+    <%
+      if (editCardSet != null) {
+    %>
+      Editing <span style="text-decoration:italic"><%=editCardSet.getName()%></span>
+    <%
+      } else {
+    %>
       Creating new card set
-    <% } %>
+    <%
+      }
+    %>
   </h2>
   <label for="cardSetName">Name:</label>
-  <input type="text" name="cardSetName" id="cardSetName"
-      value="<%= editCardSet != null ? editCardSet.getName() : "" %>" />
+  <input type="text" name="cardSetName" id="cardSetName" size="50"
+      value="<%=editCardSet != null ? StringEscapeUtils.escapeXml11(editCardSet.getName()) : ""%>" />
+  <br/>
+  <label for="cardSetDescription">Description:</label>
+  <input type="text" name="cardSetDescription" id="cardSetDescription" size="50"
+      value="<%=editCardSet != null ? StringEscapeUtils.escapeXml11(editCardSet.getDescription()) : ""%>" />
+  <br/>
+  <label for="cardSetWeight">Weight:</label>
+  <input type="text" name="cardSetWeight" id="cardSetWeight" size="4"
+      value="<%=editCardSet != null ? editCardSet.getWeight() : "1000"%>" />
   <br/>
   <label for="active">Active</label>
   <input type="checkbox" name="active" id="active"
-      <%= editCardSet != null && editCardSet.isActive() ? "checked='checked'" : "" %> />
+      <%=editCardSet != null && editCardSet.isActive() ? "checked='checked'" : ""%> />
   <br/>
   <label for="baseDeck" title="This deck is sufficient for playing the game.">Base Deck</label>
   <input type="checkbox" name="baseDeck" id="baseDeck"
-      <%= editCardSet != null && editCardSet.isBaseDeck() ? "checked='checked'" : "" %> />
+      <%=editCardSet != null && editCardSet.isBaseDeck() ? "checked='checked'" : ""%> />
   <br/>
   Available Black Cards:
   <br/>
   <select id="allBlackCards" multiple="multiple" style="height:300px">
-    <% for (BlackCard blackCard : blackCards) { %>
-      <option value="<%= blackCard.getId() %>"><%= blackCard.toString() %></option>
-    <% } %>
+    <%
+      for (PyxBlackCard blackCard : blackCards) {
+    %>
+      <option value="<%=blackCard.getId()%>">
+        <%=StringEscapeUtils.escapeXml11(blackCard.toString())%>
+      </option>
+    <%
+      }
+    %>
   </select>
   <br/>
   <input type="button" id="addBlackCards" value="Add Black Cards" />
@@ -278,21 +331,35 @@ select {
   Black Cards in Card Set:
   <br/>
   <select id="selectedBlackCards" name="selectedBlackCards" multiple="multiple">
-    <% if (editCardSet != null) { %>
-      <% for (BlackCard blackCard : editCardSet.getBlackCards()) { %>
-        <option value="<%= blackCard.getId() %>" id="bc_<%= blackCard.getId() %>">
-          <%= blackCard.toString() %>
+    <%
+      if (editCardSet != null) {
+    %>
+      <%
+        for (PyxBlackCard blackCard : editCardSet.getBlackCards()) {
+      %>
+        <option value="<%=blackCard.getId()%>" id="bc_<%=blackCard.getId()%>">
+          <%=StringEscapeUtils.escapeXml11(blackCard.toString())%>
         </option>
-      <% } %>
-    <% } %>
+      <%
+        }
+      %>
+    <%
+      }
+    %>
   </select>
   <br/>
   Available White Cards:
   <br/>
   <select id="allWhiteCards" multiple="multiple" style="height:300px">
-    <% for (WhiteCard whiteCard : whiteCards) { %>
-      <option value="<%= whiteCard.getId() %>"><%= whiteCard.toString() %></option>
-    <% } %>
+    <%
+      for (PyxWhiteCard whiteCard : whiteCards) {
+    %>
+      <option value="<%=whiteCard.getId()%>">
+        <%=StringEscapeUtils.escapeXml11(whiteCard.toString())%>
+      </option>
+    <%
+      }
+    %>
   </select>
   <br/>
   <input type="button" id="addWhiteCards" value="Add White Cards" />
@@ -301,10 +368,14 @@ select {
   White Cards in Card Set:
   <br/>
   <select id="selectedWhiteCards" name="selectedWhiteCards" multiple="multiple">
-    <% if (editCardSet != null) { %>
-      <% for (WhiteCard whiteCard : editCardSet.getWhiteCards()) { %>
+    <%
+      if (editCardSet != null) {
+    %>
+      <%
+        for (PyxWhiteCard whiteCard : editCardSet.getWhiteCards()) {
+      %>
         <option value="<%= whiteCard.getId() %>" id="wc_<%= whiteCard.getId() %>">
-          <%= whiteCard.toString() %>
+          <%= StringEscapeUtils.escapeXml11(whiteCard.toString()) %>
         </option>
       <% } %>
     <% } %>
